Move TransformCommand to a separate module
